+++ src/module.auto.test.ts
@@ -27,6 +27,7 @@ import {
   loggerErrorSpy,
   loggerFatalSpy,
   loggerInfoSpy,
+  loggerNoticeSpy,
   loggerWarnSpy,
   server,
   setDebug,
@@ -96,6 +97,8 @@ const addClusterServerColorControlSpy = jest.spyOn(MutableDevice.prototype, 'add
 const addClusterServerAutoModeThermostatSpy = jest.spyOn(MutableDevice.prototype, 'addClusterServerAutoModeThermostat');
 const addClusterServerHeatingThermostatSpy = jest.spyOn(MutableDevice.prototype, 'addClusterServerHeatingThermostat');
 const addClusterServerCoolingThermostatSpy = jest.spyOn(MutableDevice.prototype, 'addClusterServerCoolingThermostat');
+const addVacuumSpy = jest.spyOn(MutableDevice.prototype, 'addVacuum');
+const addSelectSpy = jest.spyOn(MutableDevice.prototype, 'addSelect');
 
 MatterbridgeEndpoint.logLevel = LogLevel.DEBUG; // Set the log level for MatterbridgeEndpoint to DEBUG
 
@@ -117,14 +120,14 @@ describe('Matterbridge ' + NAME, () => {
     matterbridgeVersion: '3.7.5',
     log,
     addBridgedEndpoint: jest.fn(async (pluginName: string, device: MatterbridgeEndpoint) => {
-      await addDevice(aggregator, device, MATTER_PAUSE);
+      await addDevice(aggregator, device, 1, 0);
     }),
     removeBridgedEndpoint: jest.fn(async (pluginName: string, device: MatterbridgeEndpoint) => {
-      await deleteDevice(aggregator, device, MATTER_PAUSE);
+      await deleteDevice(aggregator, device, 1, 0);
     }),
     removeAllBridgedEndpoints: jest.fn(async (pluginName: string) => {
       for (const device of aggregator.parts) {
-        await deleteDevice(aggregator, device, MATTER_PAUSE);
+        await deleteDevice(aggregator, device, 1, 0);
       }
     }),
     addVirtualEndpoint: jest.fn(async (pluginName: string, name: string, type: 'light' | 'outlet' | 'switch' | 'mounted_switch', callback: () => Promise<void>) => {
@@ -132,7 +135,7 @@ describe('Matterbridge ' + NAME, () => {
       const device = new MatterbridgeEndpoint([onOffMountedSwitch, onOffOutlet], { id: name.replaceAll(' ', '') + ':' + type })
         .createDefaultBridgedDeviceBasicInformationClusterServer(name, createUniqueId(), VendorId(0xfff1), 'Matterbridge', 'Matterbridge Virtual Device')
         .addRequiredClusterServers();
-      await addDevice(aggregator, device, MATTER_PAUSE);
+      await addDevice(aggregator, device, 1, 0);
     }),
   } as any;
 
@@ -733,6 +736,91 @@ describe('Matterbridge ' + NAME, () => {
     const endpoint = haPlatform.matterbridgeDevices.get(device.id);
     expect(endpoint).toBeDefined();
     expect(endpoint?.getChildEndpoints().length).toBe(0);
+  });
+
+  it('should call onStart and register a remote individual entity, a device with two remote entities, one normal and one split with Merge strategy', async () => {
+    const device = generateDevice(haPlatform.ha, 'TV Device');
+    const remoteIndividualEntity = generateEntity(haPlatform.ha, 'Individual remote', 'remote');
+    const remoteDeviceEntity = generateEntity(haPlatform.ha, 'Device remote entity', 'remote', device);
+    const remoteSplitEntity = generateEntity(haPlatform.ha, 'Split remote entity', 'remote', device);
+    generateState(haPlatform.ha, remoteIndividualEntity, 'on');
+    generateState(haPlatform.ha, remoteDeviceEntity, 'on');
+    generateState(haPlatform.ha, remoteSplitEntity, 'on');
+
+    haPlatform.config.splitEntities = [remoteSplitEntity.entity_id];
+    haPlatform.config.controllerStrategy = 'Merge';
+
+    await haPlatform.onStart('Test reason');
+    expect(loggerInfoSpy).toHaveBeenCalledWith(`Starting platform ${idn}${mockConfig.name}${rs}${nf}: Test reason`);
+    expect(loggerDebugSpy).toHaveBeenCalledWith(`Registering device ${dn}${remoteIndividualEntity.original_name}${db}...`);
+    expect(loggerDebugSpy).toHaveBeenCalledWith(`Registering device ${dn}${device.name}${db}...`);
+    expect(loggerDebugSpy).toHaveBeenCalledWith(`Registering device ${dn}${remoteSplitEntity.original_name}${db}...`);
+    expect(loggerInfoSpy).toHaveBeenCalledWith(`Started platform ${idn}${mockConfig.name}${rs}${nf}: Test reason`);
+    expect(mockMatterbridge.addBridgedEndpoint).toHaveBeenCalledTimes(3);
+    expect(haPlatform.matterbridgeDevices.size).toBe(3);
+    expect(haPlatform.matterbridgeDevices.has(remoteIndividualEntity.entity_id)).toBe(true);
+    expect(haPlatform.matterbridgeDevices.has(device.id)).toBe(true);
+    expect(haPlatform.matterbridgeDevices.has(remoteDeviceEntity.entity_id)).toBe(false);
+    expect(haPlatform.matterbridgeDevices.has(remoteSplitEntity.entity_id)).toBe(true);
+    expect(haPlatform.endpointNames.size).toBe(3);
+    expect(haPlatform.endpointNames.get(remoteIndividualEntity.entity_id)).toBe('');
+    expect(haPlatform.endpointNames.get(remoteDeviceEntity.entity_id)).toBe('');
+    expect(haPlatform.endpointNames.get(remoteSplitEntity.entity_id)).toBe('');
+    expect(aggregator.parts.size).toBe(3);
+    const endpoint = haPlatform.matterbridgeDevices.get(device.id);
+    expect(endpoint).toBeDefined();
+    expect(endpoint?.getChildEndpoints().length).toBe(0);
+
+    jest.clearAllMocks();
+    haPlatform.filterMessages.length = 0;
+    await haPlatform.onConfigure();
+    expect(loggerWarnSpy).not.toHaveBeenCalled();
+    expect(loggerErrorSpy).not.toHaveBeenCalled();
+    expect(loggerFatalSpy).not.toHaveBeenCalled();
+  });
+
+  it('should call onStart and register a input_select individual entity, a device with two remote entities, one normal and one split with Merge strategy', async () => {
+    const device = generateDevice(haPlatform.ha, 'Select Device');
+    const inputSelectIndividualEntity = generateEntity(haPlatform.ha, 'Individual input_select', 'input_select');
+    const selectDeviceEntity = generateEntity(haPlatform.ha, 'Device select entity', 'select', device);
+    const selectSplitEntity = generateEntity(haPlatform.ha, 'Split select entity', 'select', device);
+    generateState(haPlatform.ha, inputSelectIndividualEntity, 'Led On', { options: ['Led On', 'Led Off'], friendly_name: 'Individual input_select Led' });
+    generateState(haPlatform.ha, selectDeviceEntity, 'Led Off', { options: ['Led On', 'Led Off'], friendly_name: 'Device select entity Led' });
+    generateState(haPlatform.ha, selectSplitEntity, 'invalid', { options: ['Led On', 'Led Off'], friendly_name: 'Split select entity Led' });
+
+    haPlatform.config.splitEntities = [selectSplitEntity.entity_id];
+    haPlatform.config.controllerStrategy = 'Merge';
+
+    await setDebug(true);
+
+    await haPlatform.onStart('Test reason');
+    expect(loggerInfoSpy).toHaveBeenCalledWith(`Starting platform ${idn}${mockConfig.name}${rs}${nf}: Test reason`);
+    expect(loggerDebugSpy).toHaveBeenCalledWith(`Registering device ${dn}${inputSelectIndividualEntity.original_name}${db}...`);
+    expect(loggerDebugSpy).toHaveBeenCalledWith(`Registering device ${dn}${device.name}${db}...`);
+    expect(loggerDebugSpy).toHaveBeenCalledWith(`Registering device ${dn}${selectSplitEntity.original_name}${db}...`);
+    expect(loggerInfoSpy).toHaveBeenCalledWith(`Started platform ${idn}${mockConfig.name}${rs}${nf}: Test reason`);
+    expect(mockMatterbridge.addBridgedEndpoint).toHaveBeenCalledTimes(3);
+    expect(addSelectSpy).toHaveBeenCalledTimes(3);
+    expect(haPlatform.matterbridgeDevices.size).toBe(3);
+    expect(haPlatform.matterbridgeDevices.has(inputSelectIndividualEntity.entity_id)).toBe(true);
+    expect(haPlatform.matterbridgeDevices.has(device.id)).toBe(true);
+    expect(haPlatform.matterbridgeDevices.has(selectDeviceEntity.entity_id)).toBe(false);
+    expect(haPlatform.matterbridgeDevices.has(selectSplitEntity.entity_id)).toBe(true);
+    expect(haPlatform.endpointNames.size).toBe(3);
+    expect(haPlatform.endpointNames.get(inputSelectIndividualEntity.entity_id)).toBe('');
+    expect(haPlatform.endpointNames.get(selectDeviceEntity.entity_id)).toBe('');
+    expect(haPlatform.endpointNames.get(selectSplitEntity.entity_id)).toBe('');
+    expect(aggregator.parts.size).toBe(3);
+    const endpoint = haPlatform.matterbridgeDevices.get(device.id);
+    expect(endpoint).toBeDefined();
+    expect(endpoint?.getChildEndpoints().length).toBe(0);
+
+    jest.clearAllMocks();
+    haPlatform.filterMessages.length = 0;
+    await haPlatform.onConfigure();
+    expect(loggerWarnSpy).not.toHaveBeenCalled();
+    expect(loggerErrorSpy).not.toHaveBeenCalled();
+    expect(loggerFatalSpy).not.toHaveBeenCalled();
   });
 
   it('should call onStart and not register an individual entity, a device with two entities, one normal and one split with discardHiddenEntities enabled', async () => {